JDF-510: change the use of the scoped-context to avoid a TooManyChannelException and use the Context global for this bean to prevent also from performance issues due to massive re-creation and parallel connections for several bean instances. To reflect the correct terms 'EJBClient34' is renamed to to 'ScopedContext'.
