+++ ejb-multi-server/app-main/web/src/main/webapp/ejbinvoke.xhtml
@@ -46,8 +46,8 @@
             </h:panelGrid>
             <h:panelGrid columns="1">
                 <h:commandButton id="mainEjb" value="Invoke local MainApp" action="#{jsfController.callEJBMainLocal()}" />
-                <h:commandButton id="main34Ejb" value="Invoke local MainEjbClient34App"
-                    action="#{jsfController.callEJBClient34MainLocal()}" />
+                <h:commandButton id="mainEjbSC" value="Invoke local MainApp with scoped context"
+                    action="#{jsfController.callEJBMainScopedContextLocal()}" />
                 <h:commandButton id="oneEjb" value="Invoke remote AppOne" action="#{jsfController.callEJBAppOneRemote()}" />
                 <h:commandButton id="twoEjb" value="Invoke remote AppTwo" action="#{jsfController.callEJBAppTwoRemote()}" />
             </h:panelGrid>
